+++ handler/api/api.go
@@ -188,9 +188,9 @@ func (s Server) Handler() http.Handler {
 				acl.CheckWriteAccess(),
 			).Post("/{number}/promote", builds.HandlePromote(s.Repos, s.Builds, s.Triggerer))
 
-			// r.With(
-			// 	acl.CheckAdminAccess(),
-			// ).Post("/{number}/rollback", builds.HandleRollback(s.Repos, s.Builds, s.Triggerer))
+			r.With(
+				acl.CheckAdminAccess(),
+			).Post("/{number}/rollback", builds.HandleRollback(s.Repos, s.Builds, s.Triggerer))
 
 			r.With(
 				acl.CheckAdminAccess(),